chore: typecheck tests
