+++ packages/cli/package.json
@@ -50,7 +50,7 @@
   },
   "devDependencies": {
     "@types/which": "^2.0.1",
-    "typescript": "5.0.0-dev.20221208",
+    "typescript": "4.5.5",
     "vitest": "^0.23.4"
   },
   "peerDependencies": {
