+++ packages/cli/src/__tests__/fixtures/app/foo_2/foo_2a.ts
@@ -1,0 +1,46 @@
+
+
+interface GitDescribe {
+  tag: string;
+  count: number;
+  sha: string;
+  dirty: boolean;
+}
+
+/* @ts-expect-error @rehearsal TODO TS2451: Cannot redeclare block-scoped variable 'VERSION_PATTERN'. */
+const VERSION_PATTERN = /_(\d+\.\d+\.\d+)/;
+
+
+
+
+
+
+
+function timestamp(inSeconds = false): number {
+  return inSeconds ? Date.now() / 1000 : Date.now();
+}
+
+function parseLongDescribe(desc: string): GitDescribe {
+  const result = /^(.+)-(\d+)-g([a-f0-9]+)(?:-(dirty))?$/.exec(desc);
+
+  if (!result) {
+    throw Error(`Unable to parse ${desc} as a long description`);
+  }
+
+  const [, tag, count, sha, dirty] = result;
+
+  return {
+    tag,
+    count: parseInt(count, 10),
+    sha,
+    dirty: !!dirty,
+  };
+}
+
+function normalizeVersionString(versionString: string): string {
+  if (VERSION_PATTERN.test(versionString)) {
+    const result = VERSION_PATTERN.exec(versionString);
+    return result ? result[1] : versionString;
+  }
+  return versionString;
+}
