+++ packages/cli/src/__tests__/fixtures/app/tsconfig.json
