+++ packages/cli/src/__tests__/fixtures/app_for_migrate/src/basic/foo.js
