+++ packages/cli/src/__tests__/fixtures/app_for_migrate/src/custom_basepath/base/package.json
