+++ packages/cli/src/__tests__/fixtures/app_for_migrate/tsconfig.json
