+++ packages/cli/src/__tests__/fixtures/app_with_comments/foo.ts
